# bx-super-app-backend/spec/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Super APP",
    "description": "Super APP Services",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "UploadPhoto": { "type": "object", "properties": {} },
      "UserEdit": {
        "type": "object",
        "properties": {
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "phone": { "type": "string" }
        },
        "required": ["firstName", "lastName", "phone"]
      },
      "ChangePassword": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "new_password": { "type": "string" }
        },
        "required": ["username", "new_password"]
      },
      "EditDefaultPymeModel": {
        "type": "object",
        "properties": { "pyme_id": { "type": "string" } },
        "required": ["pyme_id"]
      },
      "UserFirstLogin": {
        "type": "object",
        "properties": { "firstLogin": { "type": "boolean" } },
        "required": ["firstLogin"]
      },
...
    }
  },
  "paths": {
...
    "/api/pyme2c/backend/v1/user": {
      "get": {
        "operationId": "UserController_getUserInfo",
        "parameters": [],
        "responses": {
          "200": { "description": "User Info" },
          "401": { "description": "Unauthorized" }
        },
        "tags": ["User"],
        "security": [{ "bearer": [] }]
      },
      "put": {
        "operationId": "UserController_editUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserEdit" }
            }
          }
        },
        "responses": {
          "200": { "description": "Edit User" },
          "401": { "description": "Unauthorized" }
        },
        "tags": ["User"],
        "security": [{ "bearer": [] }]
      }
    },
    "/api/pyme2c/backend/v1/user/upload-photo": {
      "put": {
        "operationId": "UserController_uploadPhoto",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UploadPhoto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Upload Photo" },
          "401": { "description": "Unauthorized" }
        },
        "tags": ["User"],
        "security": [{ "bearer": [] }]
      }
    },
    "/api/pyme2c/backend/v1/user/change-password": {
      "put": {
        "operationId": "UserController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePassword" }
            }
          }
        },
        "responses": {
          "200": { "description": "Change Password" },
          "401": { "description": "Unauthorized" }
        },
        "tags": ["User"],
        "security": [{ "bearer": [] }]
      }
    },
    "/api/pyme2c/backend/v1/user/edit-default-pyme": {
      "post": {
        "operationId": "UserController_editDefaultPyme",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EditDefaultPymeModel" }
            }
          }
        },
        "responses": {
          "200": { "description": "Edit Default Pyme" },
          "401": { "description": "Unauthorized" }
        },
        "tags": ["User"],
        "security": [{ "bearer": [] }]
      }
    },
    "/api/pyme2c/backend/v1/user/first-login": {
      "put": {
        "operationId": "UserController_editUserFirstLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserFirstLogin" }
            }
          }
        },
        "responses": {
          "200": { "description": "Edit User" },
          "401": { "description": "Unauthorized" }
        },
        "tags": ["User"],
        "security": [{ "bearer": [] }]
      }
    },
...
  }
}